bugs & errors fixed
