+++ src/components/initial_page/InitialPage.jsx
@@ -26,77 +26,7 @@ const style = {
 };
 
 
-const InitialPage = ({setView, setUser, setGuest, setSignedIn, setPreviewImage}) => {
-    const [open, setOpen] = useState(false);
-    const [signInEmail, setSignInEmail] = useState('');
-    const [signInPassword, setSignInPassword] = useState('');
-    const [showEmailMessage, setShowEmailMessage] = useState(false);
-    const [showInvalidEmail, setShowInvalidEmail] = useState(false);
-    const [showMissingPass, setShowMissingPass] = useState(false);
-    const [showPasswordMessage, setShowPasswordMessage] = useState(false);
-
-  const handleOpen = () => {
-    setOpen(true);
-  };
-  const handleClose = () => {
-    setOpen(false);
-  };
-  const onEmailChange = (event) => {
-    setSignInEmail(event.target.value);
-  };
-  const onPasswordChange = (event) => {
-    setSignInPassword(event.target.value);
-  };
-  // send get/post request to database with user details
-  // if login user matches, set view to trading
-  // else return message that login doesnt match any account in database
-
-  const getUser = async (signInEmail, signInPassword) => {
-      try {
-          const userData = await controllers.getUser(signInEmail, signInPassword);
-          console.log('what is this: ', userData.code)
-          if (userData.code === 'auth/invalid-email') {
-            setShowEmailMessage(true);
-          }
-          if (userData.code === 'auth/user-not-found') {
-            setShowEmailMessage(false);
-            setShowInvalidEmail(true);
-            setShowPasswordMessage(false);
-            setShowMissingPass(true);
-          }
-          if (userData.code === 'auth/missing-password') {
-            setShowEmailMessage(false);
-            setShowMissingPass(true);
-          }
-          if (userData.code === 'auth/wrong-password') {
-            setShowEmailMessage(false);
-            setShowMissingPass(false);
-            setShowInvalidEmail(false);
-            setShowPasswordMessage(true);
-          }
-          if(!userData.code){
-              setSignedIn(true);
-              setUser(userData);
-              setPreviewImage(userData.profilePic);
-              setShowEmailMessage(false);
-              setShowMissingPass(false);
-              setShowPasswordMessage(false);
-          }
-          console.log('User Data: ', userData);
-      } catch(error) {
-          setSignedIn(false);
-          console.log('error signing in: ', error);
-      }
-  }
-
-    // const guestUser = {
-    //     uid: null,
-    //     firstName: 'Guest',
-    //     lastName: null,
-    //     email: null,
-    //     totalAssets: null,
-    //     coinsOwned: null
-    // }
+const InitialPage = ({ handleOpen, handleClose, open, setView, setUser, setGuest, setSignedIn, setPreviewImage, onEmailChange, onPasswordChange, getUser, signInEmail, signInPassword, showEmailMessage, showPasswordMessage, showInvalidEmail, showMissingPass}) => {
 
     const guestLogin = async () => {
         const guestUser = await controllers.getUserById();
@@ -169,8 +99,6 @@ const InitialPage = ({setView, setUser, setGuest, setSignedIn, setPreviewImage})
       <Dialog open={open} aria-labelledby="dialog-title" sx={style}>
         <DialogTitle id="dialog-title">Register</DialogTitle>
         <Register getUser={getUser} handleClose={handleClose} />
-        {/* <Box>
-        </Box> */}
       </Dialog>
     </Box>
   );
